Accept signing key ID as alternative to key file
In automated build environments, it might be acceptable to have private
signing key material available in file form, so it should be fine to
hand over the key to Attune for signing.

However, in some scenarios, that’s not feasible. For example, a senior
package developer might be evaluating Attune for their team, and they
keep their signing key on a personal YubiKey, so they can’t export their
private key material for Attune even if they were willing to (and that’s
a big if).

Or a package maintainer might be using GPG agent forwarding, e.g.
because they’re running the Attune CLI in a container, or for
troubleshooting a multi-architecture build on a remote build node.

To support these scenarios, add a `--signing-key-id` switch to the CLI
as an alternative to `--signing-key-file`.